+++ lib/aramanim.py
@@ -4,45 +4,14 @@ from manim import TexTemplate, MathTex
 from manim import FadeIn, FadeOut, Animation, Scene
 from manim import always_redraw
 import numpy as np
-from numpy import linalg as la
 from objects import Scissors, Stopwatch
 
-# np վեկտորների համար ՛մեթոդներ՛
-def length(array):
-    return la.norm(array)
-
-def normalized_vector(array):
-    array = np.array(array)
-    return (1/length(array)) * array
-
-def rotate2angle(array, alpha):
-    array = np.array(array)
-    rotation__matrix = [
-        [np.cos(alpha), -np.sin(alpha), 0],
-        [np.sin(alpha), np.cos(alpha), 0],
-        [0, 0, 1],
-    ]
-    return np.matmul(rotation__matrix, array)
-
-def normal(array):
-    array = np.array(array)
-    return normalized_vector(rotate2angle(array, PI/2))
-
-# հայերեն լեզվով LaTeX աշխատացնելու համար
-XELATEX_preamble = r"""
-\usepackage{tikz}
-\usepackage{fontspec}
-\setmainfont{DejaVu Serif}
-\usepackage{amsfonts,amssymb,amsthm,mathtools}
-\usepackage{amsmath}
-\usepackage{upgreek}
-\usepackage{mathrsfs}
-"""
-XELATEX = TexTemplate('xelatex', '.pdf', preamble=XELATEX_preamble)
+from utilities import normal_vector
 
 # Մասերով խնդրի համար
 DEFAULT_EDGE_HEIGHT = 0.2
 DEFAULT_LABEL_FONT_SIZE = 50
+
 
 class Segment(VGroup):
     def __init__(
@@ -55,7 +24,7 @@ class Segment(VGroup):
         ):
         start = np.array(start)
         end = np.array(end)
-        normal_direction = normal(end-start)
+        normal_direction = normal_vector(end-start)
         if 'stroke_width' in kwargs:
             stroke_width = kwargs['stroke_width']
         else:
@@ -94,7 +63,8 @@ class Segment(VGroup):
         self.label.next_to(self, direction, buff)
         return self.label
 
-# մկրատի անիմացիաներ (Scissors)
+
+### մկրատի անիմացիաներ (Scissors) ###
 class CutIn(FadeIn):
     def __init__(self, mobject: Scissors, **kwargs):
         super().__init__(mobject, **kwargs)
@@ -110,6 +80,7 @@ class CutIn(FadeIn):
     def clean_up_from_scene(self, scene: Scene) -> None:
         scene.add(self.mobject)
 
+
 class CutOut(FadeOut):
     def __init__(self, mobject: Scissors, **kwargs):
         super().__init__(mobject, **kwargs)
@@ -123,7 +94,9 @@ class CutOut(FadeOut):
     def clean_up_from_scene(self, scene: Scene) -> None:
         scene.remove(self.mobject)
 
-# ժամացույցի անիմացիա
+### մկրատի անիմացիաներ (Scissors) END ###
+
+### ժամացույցի անիմացիա  ###
 class Run(Animation):
     def __init__(
         self,
@@ -211,3 +184,5 @@ class Reset(Animation):
             about_point=self.about_point
         )
         self.mobject.time.set_value(0)
+
+### ժամացույցի անիմացիա  END ###
