--- problems/problem_sample.py
@@ -1,4 +1,0 @@
-def run():
-    '''Sample problem runner'''
-
-    print('Running sample problem')
